added status handling depending on profile image
